feat(ui): widen the agent chat API surface for embedding consumers
The agent conversation carries several pieces of logic that consumers
embedding ModernAgentConversation need but cannot reach. Some symbols are
`export`ed but never surfaced through a published subpath's barrel chain, so
they are unreachable for anything resolving `@vertesia/ui/*` against the
package's `exports` map. Others are module-private inside a large `.tsx`.

Surface through `@vertesia/ui/features`:

- `useDocumentPanel`, `useArtifacts` (plus `ArtifactTreeNode`)
- the `workstreams.ts` helpers and the `WorkstreamInfo` type

Extract from `ModernAgentConversation.tsx` into exported modules:

- `deriveWorkstreamsFromMessages` and the `getActiveWorkstreams` merge helpers
  into `chat/deriveWorkstreams.ts` — a sibling module rather than
  `workstreams.ts`, which `ModernAgentOutput/utils.ts` already imports from;
  merging them would introduce an import cycle
- `deriveArtifactRefreshKey` into `chat/hooks/useArtifacts.ts`
- the active-workstreams polling effect into
  `chat/hooks/useActiveWorkstreams.ts`, which now owns the poll interval, the
  unavailability latch and the warn-once ref; callers pass only the conditions
  they own

Also backport the shared `OverflowTabsBar` / `OverflowMoreMenu` /
`useOverflowTabs` component from #1896 on main, so the measure-and-collapse tab
bar is reusable from `@vertesia/ui/core` instead of living inside
`AgentRightPanel`.

The moved logic is byte-identical to the originals and behaviour is unchanged,
with one exception: a change of `initialHistoryStatus` between two
non-`loading` values no longer restarts the workstream poll and its redundant
immediate fetch.
