+++ packages/ui/src/features/agent/chat/ModernAgentOutput/WorkstreamTabs.tsx
@@ -1,8 +1,7 @@
 import type { AgentMessage } from '@vertesia/common';
-import { cn, DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@vertesia/ui/core';
+import { cn, DropdownMenuItem, OverflowMoreMenu, useOverflowTabs } from '@vertesia/ui/core';
 import { i18nInstance, NAMESPACE, useUITranslation } from '@vertesia/ui/i18n';
 import { CheckCircle, ChevronDown, Clock } from 'lucide-react';
-import { useEffect, useLayoutEffect, useRef, useState } from 'react';
 import { getWorkstreamId } from './utils';
 
 interface WorkstreamTabsProps {
@@ -19,7 +18,6 @@ const TAB_ITEM_BASE =
     'flex items-center gap-1.5 px-2 py-1 text-xs font-medium whitespace-nowrap transition-colors border-b-2 shrink-0 cursor-pointer';
 const TAB_ITEM_INACTIVE = 'border-transparent text-muted hover:bg-muted';
 const TAB_ITEM_ACTIVE = 'border-info bg-info text-info';
-const TAB_GAP_PX = 4; // matches the `gap-1` between tab items
 
 // Shorten long workstream names for the tab row.
 function truncateName(name: string) {
@@ -50,94 +48,39 @@ function WorkstreamMoreMenu({
     count,
     completionStatus,
 }: WorkstreamMoreMenuProps) {
-    const [open, setOpen] = useState(false);
-    const closeTimer = useRef<ReturnType<typeof setTimeout> | null>(null);
-    const openedByHover = useRef(false);
-
-    const cancelClose = () => {
-        if (closeTimer.current) {
-            clearTimeout(closeTimer.current);
-            closeTimer.current = null;
-        }
-    };
-    const openOnHover = () => {
-        cancelClose();
-        openedByHover.current = true;
-        setOpen(true);
-    };
-    // Delay the close so the pointer can travel across the gap onto the menu.
-    const closeAfterDelay = () => {
-        cancelClose();
-        closeTimer.current = setTimeout(() => setOpen(false), 150);
-    };
-
-    // Clear any pending close timer on unmount.
-    useEffect(() => () => clearTimeout(closeTimer.current ?? undefined), []);
-
     return (
-        // Non-modal: a modal menu sets body `pointer-events: none` while open, which makes
-        // the hover open/close flap. It still closes on outside-click / Escape.
-        <DropdownMenu
-            modal={false}
-            open={open}
-            onOpenChange={(next) => {
-                // Fired by click / keyboard / dismiss (not by hover); keep our state in sync.
-                cancelClose();
-                if (next) openedByHover.current = false;
-                setOpen(next);
-            }}
+        <OverflowMoreMenu
+            label={label}
+            triggerClassName={cn(TAB_ITEM_BASE, active ? TAB_ITEM_ACTIVE : TAB_ITEM_INACTIVE)}
+            icon={<ChevronDown className="ms-0.5 size-3.5" />}
         >
-            <DropdownMenuTrigger asChild>
-                {/* Tab-bar primitive: raw button is the menu trigger (asChild). */}
-                <button
-                    type="button"
-                    onMouseEnter={openOnHover}
-                    onMouseLeave={closeAfterDelay}
-                    className={cn(TAB_ITEM_BASE, active ? TAB_ITEM_ACTIVE : TAB_ITEM_INACTIVE)}
-                >
-                    {label}
-                    <ChevronDown className="ms-0.5 size-3.5" />
-                </button>
-            </DropdownMenuTrigger>
-            <DropdownMenuContent
-                align="end"
-                className="w-max"
-                onMouseEnter={cancelClose}
-                onMouseLeave={closeAfterDelay}
-                onCloseAutoFocus={(e) => {
-                    // Keep hover-opens from returning the focus ring to the trigger.
-                    if (openedByHover.current) e.preventDefault();
-                }}
-            >
-                {items.map(([id, name]) => {
-                    // biome-ignore lint/style/noNonNullAssertion: guarded by count?.has(id).
-                    const showCount = count?.has(id) && count.get(id)! > 0;
-                    return (
-                        <DropdownMenuItem
-                            key={id}
-                            onClick={() => onSelect(id)}
-                            className={cn('flex items-center gap-2', id === current && 'text-info')}
-                        >
-                            <span className="truncate">{name}</span>
-                            {showCount && (
-                                <span className="ms-auto inline-flex items-center gap-1">
-                                    <span className="inline-flex items-center justify-center rounded-full bg-muted px-1.5 text-[10px] text-muted">
-                                        {count?.get(id)}
-                                    </span>
-                                    {completionStatus &&
-                                        id !== 'all' &&
-                                        (completionStatus.get(id) ? (
-                                            <CheckCircle className="size-3 text-success" />
-                                        ) : (
-                                            <Clock className="size-3 text-attention" />
-                                        ))}
+            {items.map(([id, name]) => {
+                const showCount = (count?.get(id) ?? 0) > 0;
+                return (
+                    <DropdownMenuItem
+                        key={id}
+                        onClick={() => onSelect(id)}
+                        className={cn('flex items-center gap-2', id === current && 'text-info')}
+                    >
+                        <span className="truncate">{name}</span>
+                        {showCount && (
+                            <span className="ms-auto inline-flex items-center gap-1">
+                                <span className="inline-flex items-center justify-center rounded-full bg-muted px-1.5 text-[10px] text-muted">
+                                    {count?.get(id)}
                                 </span>
-                            )}
-                        </DropdownMenuItem>
-                    );
-                })}
-            </DropdownMenuContent>
-        </DropdownMenu>
+                                {completionStatus &&
+                                    id !== 'all' &&
+                                    (completionStatus.get(id) ? (
+                                        <CheckCircle className="size-3 text-success" />
+                                    ) : (
+                                        <Clock className="size-3 text-attention" />
+                                    ))}
+                            </span>
+                        )}
+                    </DropdownMenuItem>
+                );
+            })}
+        </OverflowMoreMenu>
     );
 }
 
@@ -162,89 +105,16 @@ function WorkstreamOverflowBar({
     completionStatus,
 }: WorkstreamOverflowBarProps) {
     const { t } = useUITranslation();
-    const containerRef = useRef<HTMLDivElement | null>(null);
-    const itemRefs = useRef<Array<HTMLButtonElement | null>>([]);
-    const moreRef = useRef<HTMLButtonElement | null>(null);
-    // `count` leading tabs are shown. When `promote` is set, the active tab is pulled
-    // into the last slot (before More) and `count` counts the tabs shown before it.
-    const [layout, setLayout] = useState<{ count: number; promote: boolean }>({
-        count: entries.length,
-        promote: false,
-    });
-
-    const recompute = () => {
-        const container = containerRef.current;
-        if (!container) return;
-        const containerWidth = container.clientWidth;
-        const widths = entries.map((_, i) => itemRefs.current[i]?.offsetWidth ?? 0);
-        const totalAll = widths.reduce((sum, w) => sum + w, 0) + TAB_GAP_PX * Math.max(0, entries.length - 1);
-
-        // How many tabs (in order, optionally skipping one) fit within `available`.
-        const fitCount = (available: number, skipIndex: number) => {
-            let used = 0;
-            let fitted = 0;
-            for (let i = 0; i < entries.length; i++) {
-                if (i === skipIndex) continue;
-                const cand = used + (fitted > 0 ? TAB_GAP_PX : 0) + widths[i];
-                if (cand > available) break;
-                used = cand;
-                fitted += 1;
-            }
-            return fitted;
-        };
-
-        let next: { count: number; promote: boolean };
-        if (totalAll <= containerWidth) {
-            next = { count: entries.length, promote: false };
-        } else {
-            const moreWidth = moreRef.current?.offsetWidth ?? 0;
-            const naturalCount = Math.max(1, fitCount(containerWidth - moreWidth - TAB_GAP_PX, -1));
-            const activeIndex = entries.findIndex(([id]) => id === activeWorkstream);
-            if (activeIndex < 0 || activeIndex < naturalCount) {
-                next = { count: naturalCount, promote: false };
-            } else {
-                // Active tab overflowed: reserve its slot at the end, fit leading tabs before it.
-                const leadAvailable = containerWidth - moreWidth - widths[activeIndex] - TAB_GAP_PX * 2;
-                next = { count: fitCount(leadAvailable, activeIndex), promote: true };
-            }
-        }
-        setLayout((prev) => (prev.count === next.count && prev.promote === next.promote ? prev : next));
-    };
-
-    // Re-measure after every render (tab/label changes) and on container resize.
-    const recomputeRef = useRef(recompute);
-    recomputeRef.current = recompute;
-    useLayoutEffect(() => {
-        recomputeRef.current();
-    });
-    useEffect(() => {
-        const container = containerRef.current;
-        if (!container || typeof ResizeObserver === 'undefined') return;
-        const observer = new ResizeObserver(() => recomputeRef.current());
-        observer.observe(container);
-        return () => observer.disconnect();
-    }, []);
-
-    const activeIndex = entries.findIndex(([id]) => id === activeWorkstream);
-    let visible: WorkstreamEntry[];
-    let overflow: WorkstreamEntry[];
-    if (layout.promote && activeIndex >= 0) {
-        // Pull the active tab into the last visible slot; the tab it displaces overflows.
-        const others = entries.filter((_, i) => i !== activeIndex);
-        visible = [...others.slice(0, layout.count), entries[activeIndex]];
-        overflow = others.slice(layout.count);
-    } else {
-        visible = entries.slice(0, layout.count);
-        overflow = entries.slice(layout.count);
-    }
-    const activeInOverflow = overflow.some(([id]) => id === activeWorkstream);
+    const { containerRef, setItemRef, setMoreRef, visible, overflow, activeInOverflow } = useOverflowTabs(
+        entries,
+        ([id]) => id === activeWorkstream,
+    );
     const moreLabel = t('agent.moreTabs');
 
     // Inner tab content (name + count badge + completion icon), shared by the hidden
     // measurement row and the visible row so their measured widths match.
     const renderTabInner = (id: string, name: string, isActive: boolean) => {
-        // biome-ignore lint/style/noNonNullAssertion: guarded by count?.has(id).
-        const showCount = count?.has(id) && count.get(id)! > 0;
+        const showCount = (count?.get(id) ?? 0) > 0;
         return (
             <>
                 {truncateName(name)}
@@ -282,15 +152,13 @@ function WorkstreamOverflowBar({
                         type="button"
                         key={id}
                         tabIndex={-1}
-                        ref={(el) => {
-                            itemRefs.current[i] = el;
-                        }}
+                        ref={setItemRef(i)}
                         className={cn(TAB_ITEM_BASE, TAB_ITEM_INACTIVE)}
                     >
                         {renderTabInner(id, name, false)}
                     </button>
                 ))}
-                <button type="button" tabIndex={-1} ref={moreRef} className={cn(TAB_ITEM_BASE, TAB_ITEM_INACTIVE)}>
+                <button type="button" tabIndex={-1} ref={setMoreRef} className={cn(TAB_ITEM_BASE, TAB_ITEM_INACTIVE)}>
                     {moreLabel}
                     <ChevronDown className="ms-0.5 size-3.5" />
                 </button>
